backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[project]
name = "nexustext-ai"
version = "7.0.0"
description = "Global Multimodal Autonomous Text Analyzer"
requires-python = ">=3.11"
dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",
    # LLM
    "anthropic>=0.40.0",
    "openai>=1.50.0",
    "google-cloud-aiplatform>=1.70.0",
    "langchain>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-google-vertexai>=2.0.0",
    # Text mining
    "scikit-learn>=1.5.0",
    "hdbscan>=0.8.33",
    "umap-learn>=0.5.6",
    "sentence-transformers>=3.0.0",
    "bertopic>=0.16.0",
    "spacy>=3.7.0",
    "networkx>=3.3",
    "python-louvain>=0.16",
    # NLP / Morphological analysis
    "fugashi>=1.3.0",
    "unidic-lite>=1.0.8",
    "jieba>=0.42.1",
    # Data processing
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "openpyxl>=3.1.0",
    "python-docx>=1.1.0",
    "pdfplumber>=0.11.0",
    "chardet>=5.2.0",
    # Report generation
    "python-pptx>=0.6.23",
    "reportlab>=4.2.0",
    "jinja2>=3.1.0",
    # Security
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "presidio-analyzer>=2.2.0",
    "presidio-anonymizer>=2.2.0",
    # Database
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "redis>=5.0.0",
    # Cloud SDKs
    "boto3>=1.34.0",
    "azure-identity>=1.17.0",
    "azure-keyvault-secrets>=4.8.0",
    "google-cloud-secret-manager>=2.20.0",
    # Observability
    "structlog>=24.0.0",
    "opentelemetry-api>=1.25.0",
    "opentelemetry-sdk>=1.25.0",
    "opentelemetry-instrumentation-fastapi>=0.46b0",
    "opentelemetry-instrumentation-httpx>=0.46b0",
    "opentelemetry-exporter-otlp>=1.25.0",
    "prometheus-client>=0.20.0",
    # Config
    "pydantic-settings>=2.4.0",
    "pydantic>=2.8.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
    "pre-commit>=3.8.0",
]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM"]
ignore = [
    "B005",   # strip() with multi-char strings (LLM output parsing pattern)
    "B008",   # function call in defaults (FastAPI dependency injection pattern)
    "B905",   # zip() strict parameter
    "UP042",  # StrEnum (keep str+Enum for 3.11 compat)
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
strict = true
